test(ai-proxy): 5 mutation-quick-wins — MSI 48% -> 58% on AIProxyService
Addresses the pilot-run escaped mutations from
docs/kb/runbooks/infection-setup-plan.md §2. Target: 90%. Achieved
+10pp; remaining escapes are HTTP-config mutations (maxTokens
default, Http timeout, header keys) — those need request-
introspection tests (separate pass).

Added:
- test_chat_logs_execution_time_in_milliseconds_not_seconds:
  usleep(50ms) proves round($t * 1000) produces >= 40ms. Kills the
  *1000 drop and RoundingFamily mutations that would collapse to 0.
- test_chat_logs_usage_record_contains_each_documented_key:
  per-key === assertion on tenant/input/output/total/model. Kills
  ArrayItem / ArrayItemRemoval mutations in the logUsage payload.
- test_usage_stats_period_arms_resolve_correct_since_window:
  data-provider with 8 rows covering hour/day/week/month boundaries
  from both sides. Kills any match-arm removal.
- test_usage_stats_empty_window_returns_strict_zero_integers:
  === 0 on each key + assertIsInt, killing CastInt / Increment /
  Decrement on integer defaults.
- test_default_system_prompt_is_reachable_by_a_subclass:
  anonymous subclass exposes the protected method; kills the
  protected->private visibility mutation.

Source fix (real contract tightening, not a test-only change):
- AIProxyService::getUsageStats now casts avg_execution_time_ms to
  (int). The method documents "integer milliseconds" and callers
  rely on that; round() returns float, so DB-averages surfaced as
  0.0 instead of 0. Both unit tests and downstream consumers now
  see a consistent int.

Validation:
- AIProxyServiceTest: 20 tests / 58 assertions, all green.
- HavunCore Unit suite: 559 tests, all green.
- Infection filter=AIProxyService: MSI 48% -> 58% (Mutation Code
  Coverage 100%).

+++ app/Services/AIProxyService.php
@@ -169,7 +169,7 @@ class AIProxyService
             'total_input_tokens' => (int) ($stats->total_input_tokens ?? 0),
             'total_output_tokens' => (int) ($stats->total_output_tokens ?? 0),
             'total_tokens' => (int) ($stats->total_tokens ?? 0),
-            'avg_execution_time_ms' => round($stats->avg_execution_time_ms ?? 0),
+            'avg_execution_time_ms' => (int) round($stats->avg_execution_time_ms ?? 0),
         ];
     }
